+++ supertokens_python/recipe/thirdpartyemailpassword/asyncio/__init__.py
@@ -12,11 +12,13 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
+from typing import Any, Dict, Union
+
 from supertokens_python.recipe.thirdpartyemailpassword.recipe import \
     ThirdPartyEmailPasswordRecipe
 
 
-async def create_email_verification_token(user_id: str, user_context=None):
+async def create_email_verification_token(user_id: str, user_context: Union[None, Dict[str, Any]] = None):
     if user_context is None:
         user_context = {}
     email = await ThirdPartyEmailPasswordRecipe.get_instance().get_email_for_user_id(user_id, user_context)
@@ -24,14 +26,14 @@ async def create_email_verification_token(user_id: str, user_context=None):
         user_id, email, user_context)
 
 
-async def verify_email_using_token(token: str, user_context=None):
+async def verify_email_using_token(token: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     return ThirdPartyEmailPasswordRecipe.get_instance().email_verification_recipe.recipe_implementation.verify_email_using_token(
         token, user_context)
 
 
-async def is_email_verified(user_id: str, user_context=None):
+async def is_email_verified(user_id: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     email = await ThirdPartyEmailPasswordRecipe.get_instance().get_email_for_user_id(user_id, user_context)
@@ -39,7 +41,7 @@ async def is_email_verified(user_id: str, user_context=None):
         user_id, email, user_context)
 
 
-async def unverify_email(user_id: str, user_context=None):
+async def unverify_email(user_id: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     email = await ThirdPartyEmailPasswordRecipe.get_instance().get_email_for_user_id(user_id, user_context)
@@ -47,7 +49,7 @@ async def unverify_email(user_id: str, user_context=None):
         user_id, email, user_context)
 
 
-async def revoke_email_verification_tokens(user_id: str, user_context=None):
+async def revoke_email_verification_tokens(user_id: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     email = await ThirdPartyEmailPasswordRecipe.get_instance().get_email_for_user_id(user_id, user_context)
@@ -55,65 +57,56 @@ async def revoke_email_verification_tokens(user_id: str, user_context=None):
         user_id, email, user_context)
 
 
-async def get_user_by_id(user_id: str, user_context=None):
+async def get_user_by_id(user_id: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.get_user_by_id(user_id, user_context)
 
 
-async def get_user_by_third_party_info(third_party_id: str, third_party_user_id: str, user_context=None):
+async def get_user_by_third_party_info(third_party_id: str, third_party_user_id: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
-    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.get_user_by_third_party_info(
-        third_party_id, third_party_user_id, user_context)
+    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.get_user_by_thirdparty_info(third_party_id, third_party_user_id, user_context)
 
 
-async def sign_in_up(third_party_id: str, third_party_user_id: str, email: str, email_verified: bool, user_context=None):
+async def sign_in_up(third_party_id: str, third_party_user_id: str, email: str, email_verified: bool, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
-    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.sign_in_up(third_party_id,
-                                                                                               third_party_user_id,
-                                                                                               email, email_verified,
-                                                                                               user_context)
+    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.sign_in_up(third_party_id, third_party_user_id,email, email_verified,user_context)
 
 
-async def create_reset_password_token(user_id: str, user_context=None):
+async def create_reset_password_token(user_id: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.create_reset_password_token(user_id,
                                                                                                                 user_context)
 
 
-async def reset_password_using_token(token: str, new_password: str, user_context=None):
+async def reset_password_using_token(token: str, new_password: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
-    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.reset_password_using_token(token,
-                                                                                                               new_password,
-                                                                                                               user_context)
+    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.reset_password_using_token(token,new_password, user_context)
 
 
-async def sign_in(email: str, password: str, user_context=None):
+async def sign_in(email: str, password: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.sign_in(email, password, user_context)
 
 
-async def sign_up(email: str, password: str, user_context=None):
+async def sign_up(email: str, password: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.sign_up(email, password, user_context)
 
 
-async def update_email_or_password(user_id: str, email: str = None, password: str = None, user_context=None):
+async def update_email_or_password(user_id: str, email: Union[None, str] = None, password: Union[None, str] = None, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
-    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.update_email_or_password(user_id,
-                                                                                                             user_context,
-                                                                                                             email,
-                                                                                                             password)
+    return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.update_email_or_password(user_id,email,password, user_context)
 
 
-async def get_users_by_email(email: str, user_context=None):
+async def get_users_by_email(email: str, user_context: Union[None, Dict[str, Any]]=None):
     if user_context is None:
         user_context = {}
     return await ThirdPartyEmailPasswordRecipe.get_instance().recipe_implementation.get_users_by_email(email, user_context)