+++ supertokens_python/recipe/thirdpartyemailpassword/interfaces.py
@@ -1,5 +1,5 @@
 from abc import ABC, abstractmethod
-from typing import List, Union
+from typing import Any, Dict, List, Union
 
 from supertokens_python.recipe.emailpassword.interfaces import \
     APIOptions as EmailPasswordApiOptions
@@ -14,8 +14,8 @@ from supertokens_python.recipe.thirdparty.interfaces import \
 from supertokens_python.recipe.thirdparty.interfaces import (
     AuthorisationUrlGetResponse, SignInUpPostResponse, SignInUpResult)
 from supertokens_python.recipe.thirdparty.provider import Provider
-from supertokens_python.recipe.thirdpartyemailpassword.types import (
-    User, UsersResponse)
+
+from .types import User
 
 
 class RecipeInterface(ABC):
@@ -23,42 +23,42 @@ class RecipeInterface(ABC):
         pass
 
     @abstractmethod
-    async def get_user_by_id(self, user_id: str, user_context: any) -> Union[User, None]:
+    async def get_user_by_id(self, user_id: str, user_context: Dict[str, Any]) -> Union[User, None]:
         pass
 
     @abstractmethod
-    async def get_users_by_email(self, email: str, user_context: any) -> List[User]:
+    async def get_users_by_email(self, email: str, user_context: Dict[str, Any]) -> List[User]:
         pass
 
     @abstractmethod
     async def get_user_by_thirdparty_info(self, third_party_id: str,
-                                          third_party_user_id: str, user_context: any) -> Union[User, None]:
+                                          third_party_user_id: str, user_context: Dict[str, Any]) -> Union[User, None]:
         pass
 
     @abstractmethod
     async def sign_in_up(self, third_party_id: str, third_party_user_id: str, email: str,
-                         email_verified: bool, user_context: any) -> SignInUpResult:
+                         email_verified: bool, user_context: Dict[str, Any]) -> SignInUpResult:
         pass
 
     @abstractmethod
-    async def sign_in(self, email: str, password: str, user_context: any) -> SignInResult:
+    async def sign_in(self, email: str, password: str, user_context: Dict[str, Any]) -> SignInResult:
         pass
 
     @abstractmethod
-    async def sign_up(self, email: str, password: str, user_context: any) -> SignUpResult:
+    async def sign_up(self, email: str, password: str, user_context: Dict[str, Any]) -> SignUpResult:
         pass
 
     @abstractmethod
-    async def create_reset_password_token(self, user_id: str, user_context: any) -> CreateResetPasswordResult:
+    async def create_reset_password_token(self, user_id: str, user_context: Dict[str, Any]) -> CreateResetPasswordResult:
         pass
 
     @abstractmethod
-    async def reset_password_using_token(self, token: str, new_password: str, user_context: any) -> ResetPasswordUsingTokenResult:
+    async def reset_password_using_token(self, token: str, new_password: str, user_context: Dict[str, Any]) -> ResetPasswordUsingTokenResult:
         pass
 
     @abstractmethod
-    async def update_email_or_password(self, user_id: str, user_context: any, email: str = None,
-                                       password: str = None) -> UpdateEmailOrPasswordResult:
+    async def update_email_or_password(self, user_id: str, email: Union[str, None],
+                                       password: Union[str, None], user_context: Dict[str, Any]) -> UpdateEmailOrPasswordResult:
         pass
 
 
@@ -75,40 +75,39 @@ class APIInterface(ABC):
 
     @abstractmethod
     async def authorisation_url_get(self, provider: Provider,
-                                    api_options: ThirdPartyApiOptions, user_context: any) -> AuthorisationUrlGetResponse:
+                                    api_options: ThirdPartyApiOptions, user_context: Dict[str, Any]) -> AuthorisationUrlGetResponse:
         pass
 
     @abstractmethod
-    async def thirdparty_sign_in_up_post(self, provider: Provider, code: str, redirect_uri: str, client_id: Union[str, None],
-                                         auth_code_response: Union[str, None],
-                                         api_options: ThirdPartyApiOptions, user_context: any) -> SignInUpPostResponse:
+    async def thirdparty_sign_in_up_post(self, provider: Provider, code: str, redirect_uri: str, client_id: Union[str, None],auth_code_response: Union[Dict[str, Any], None],
+                                         api_options: ThirdPartyApiOptions, user_context: Dict[str, Any]) -> SignInUpPostResponse:
         pass
 
     @abstractmethod
     async def emailpassword_sign_in_post(self, form_fields: List[FormField],
-                                         api_options: EmailPasswordApiOptions, user_context: any) -> SignInPostResponse:
+                                         api_options: EmailPasswordApiOptions, user_context: Dict[str, Any]) -> SignInPostResponse:
         pass
 
     @abstractmethod
     async def emailpassword_sign_up_post(self, form_fields: List[FormField],
-                                         api_options: EmailPasswordApiOptions, user_context: any) -> SignUpPostResponse:
+                                         api_options: EmailPasswordApiOptions, user_context: Dict[str, Any]) -> SignUpPostResponse:
         pass
 
     @abstractmethod
-    async def email_exists_get(self, email: str, api_options: EmailPasswordApiOptions, user_context: any) -> EmailExistsGetResponse:
+    async def email_exists_get(self, email: str, api_options: EmailPasswordApiOptions, user_context: Dict[str, Any]) -> EmailExistsGetResponse:
         pass
 
     @abstractmethod
     async def generate_password_reset_token_post(self, form_fields: List[FormField],
-                                                 api_options: EmailPasswordApiOptions, user_context: any) -> GeneratePasswordResetTokenPostResponse:
+                                                 api_options: EmailPasswordApiOptions, user_context: Dict[str, Any]) -> GeneratePasswordResetTokenPostResponse:
         pass
 
     @abstractmethod
-    async def password_reset_post(self, token: str, form_fields: List[FormField],
-                                  api_options: EmailPasswordApiOptions, user_context: any) -> PasswordResetPostResponse:
+    async def password_reset_post(self, form_fields: List[FormField], token: str,
+                                  api_options: EmailPasswordApiOptions, user_context: Dict[str, Any]) -> PasswordResetPostResponse:
         pass
 
     @abstractmethod
     async def apple_redirect_handler_post(self, code: str, state: str,
-                                          api_options: ThirdPartyApiOptions, user_context: any):
+                                          api_options: ThirdPartyApiOptions, user_context: Dict[str, Any]):
         pass